MXS-4474: Add test case
The test reproduces the problem where session commands caused excessive
messages to be generated as well as the case where KILL commands caused
the whole MaxScale process to stall. It also verifies that the fixes to
the problem work and that they complete in a reasonable time (around 4
seconds on my laptop).