MXS-3972 The rpl state is saved to file for information only
The rpl state as a file is strictly speaking not necessary
as of this change. However, for convenience it is used for
monitoring, testing and when manually setting up pinloki,
allowing a replica to request a future GTID. When that
happens, the reader will wait for the GTID (or a later GTID)
to appear in the file before starting to read and generate events.