MXS-4430 Remove `name` param in `openEtlConn`
The connetion names stored in the meta field are only used for UX matter.
Though, it improves the UX a bit, asking the users to fill in extra
inputs for defining the connection name seems to be clumsy. Having an
input to define ETL task name would be more reasonable.