MXS-4430 Make ETL task name editable
This renders an "edit" button next to the name of the ETL task.

This also extends `page-wrapper` to have slots as alternatives for
portal slots. This is needed to make the content of the slot to grow
to fill available space as `portal-vue` doesn't provide ability to
add classname to portal wrapper element.