MXS-4430 Handling stage actions separately
Handling all stages action in `etl-stage-ctr` component causes complexity
in checking stage completion and edge cases such as when the page is reloaded.
Another reason for the changes is that according to the UI/UX design,
there are no `Previous` and `Next` buttons, the button is labeled explicitly.
To navigate to the previous stage, the users will use the sidebar tab navigation
buttons.