MXS-2709: Fix fractional TIME types
The TIME types that had a fractional part could not be inserted and
attempting to insert one would fail with a 'Datetime field overflow'
error. For the zero date value (i.e. 0000-00-00 00:00:0), the insertion
failed with a different error even if MariaDB would accept
it. Converting the SQL type into a SQL_VARCHAR fixes both of these
problems.