MXS-4468: Test ETL with most MariaDB data types
The test check that the individual values transfer over correctly and
that the same also works when done in parallel for all the tables.

Added timeouts to the ETL operation as otherwise the tests might take
too long to fail.

Also turned off log_debug for the test as it's a little bit too verbose
for this case. If the ETL fails, the result is printed out by the tests
so for initial investigation that should be enough.