MXS-4430 Make etl actions reusable
Actions for calling `prepare` and `start` endpoints can be
made reusable.

This also removes persistent field `migration_script` as it's not
wise to persist migration script. `meta.async_query_id` can be
used to refetch the script. Once the ETL task is done, the users
won't be able to see the migration script if there is a new
task being created or connections are expired, closed.