MXS-4481 Do not exlude DUAL in a CREATE TABLE statement
DUAL is an implicitly existing table https://mariadb.com/kb/en/dual/.
The embedded parser excludes DUAL if it occurs in a SELECT, which
was the reason why qc_sqlite explicitly checks for that name and
excludes it. However, that causes the parsing of a "CREATE TABLE DUAL..."
statement to crash, as it is assumed the name will be present.
With this change a "CREATE TABLE DUAL..." statement will no longer
cause MaxScale to crash but the statement itself will still be
rejected by the server.