Use std::string_view in mxb::strtok
This removes the redundant copy and in-place modification of the input
string.

Added a hand-written test case for both the old version and the new
version to make sure they both work as expected. Also added a
deterministic "random" test to make sure the outputs are identical for
both the new and the old version.