MXS-4482: Add docker to the MaxScale VM
This is the fastest method to test the PostgreSQL ETL. Otherwise, the
test would require a normal installation on either the MaxScale VM or on
a separate VM with the associated test code to start it up. Given that
it is needed only for a tiny subset of tests, there's no real benefit to
going that far to test it, at least at this point.