MXS-4482: Delete connections after ETL
The run_etl() function didn't delete them which left stale the
connections around for the duratio of the test. This wasn't the
intention and they should be deleted as the last step.

Replaced the concat function with mxb::cat now that it's added to
maxbase. Also renamed some variables that shadowed member function
names.