MXS-4482: Add mxt::Docker helper
The class takes care of starting and stopping docker containers on the
first MaxScale VM.

Changed the TestConnections::run_test to catch uncaught exceptions and
treat them as test failures. This will make sure that cleanup is done
even if the test throws an exception, whether done intentionally or
accidentally.