MXS-4320 Allow explicit formatting of maxctrl output
In an array listing maxctrl fields, it is now possible to specify
'formatter' in addition to 'name', 'path' and 'description'.

For instance:

const list_users_fields = [
  ...
  {
    name: "Last Login",
    path: "attributes.last_login",
    description: "The last time the user logged in",
    formatter: dateToLocaleString
  }
];

If provided, it is expected to be a function taking one argument
that returns a string. The function will be used when generating
the output of maxctrl.

Intended to be used when a value returned by MaxScale should be
formatted in some specfic way before being shown to the user.

common.js now also exports a function 'dateToLocaleString' that
formats a value representing a date accoring to the current
language, locale, etc.