Update Connector-C to 3.3.3
The newer versions have support for semi-sync replication which
(hopefully) should make it automatically work with the binlogrouter. The
CONC-592 fix also might end up being helpful for the binlogrouter as it
allows the host and port to be set.

It'll also contain a fix for CONC-588 which is used for the proxy header
handling. This might require changes in MaxScale as the information
might now be duplicated even if it isn't needed anymore.

One major improvement is that C/C supposedly now reports error strings
for errno when something fails due to a system error.