MXS-4430 Add `etl-create-mode-input`
The input is a dropdown which the users can select one of the following
value `normal, ignore, replace`.

This adds a question mark icon to use for document refererence.

This also removes the `filter` input as it makes the UI looks a bit clumsy
after adding the `create_mode` input. If there are requests to add
this functionality, it can be added back later.