MXS-4430 Relocate `etl-logs` in `etl-conns-stage`
According to the UI/UX design the log container is placed on the
right side while the source and destination are all placed on the
left side. But it doesn't look reasonable as the source section
actually has more inputs than it has in the design. So this adjusts
the layout to make the UI looks balance.