MXS-2709: Show ETL progress
A GET on the query endpoint of the ETL now shows the same JSON object as
the final result. Currently this is not that useful as the only useful
thing in it are the durations and any errors that might have happened
but have not yet caused the ETL to stop.