MXS-4490 Adjust failed reconnection message
If the server returns "lost connection to server during query", a dialog
is shown to let the users reconnect or disconnect it.  For some reason,
if it fails to reconnect, the active query tab has no connection bound
to it, so it's unusable. However, other tabs are still alive, so for
the sake of simplicity, this adds message to tell the user to delete the
active query tab.