MXS-4482: Move common code into etl_common.hh
Added code for processing the ETL test cases. They are identical when it
comes to the call signatures so a common test runner function can be
used. To make the declaration of test cases easier, a macro was added.

Also added explicit timeouts to the test. This way the connections are
kept alive only for the duration of the test timeout which should speed
up failures in case they ever happen.