MXS-4491 Expose details of Session DCBs
A DCB can now return a mxb::Json containing information
of where it uses memory. That information is now used
when reporting the memory usage of a session.

In practice this means that what was earlier reported as
a single value 'connections_buffers' is now reported as
an object containing an individual object for each DCB
of the session; one for the client and one for each backend.
To make it simple to see the total usage, there is also a
key "!sum" containing the sum. It is called "!sum" to make
is sort first so that its position is not dependent upon
the names of the servers.

The top-level key "total" is now also renamed to "!sum;
"sum" because it does not show the total usage but the
sum of the exposed usage and "!sum" to make it the first
item, thus easily spotted.