MXS-4491 Expose session's i/o-activity
'Session::m_io_activity' is a 30s sliding window of the i/o
activity of a session. Each slot contains a number indicating
how many epoll events were handled for the session in question
during a second.

The function 'Session::io_activity()' simply returns the sum of
the entries in that array. It is used during rebalancing, when
deciding what session to move to another routing worker.

It's now exposed by 'maxctrl show session' and provides information
about the "busyness" of a session.