MXS-4430 Initialize blank worksheet
Initialize workspace page with a blank worksheet which is a worksheet
that has a null value on both `active_query_tab_id` and `active_etl_task_id`.

The `wke-conn-man` is now obsolete because the new UI/UX has the blank worksheet
for managing connections. Therefore, it's now moved to `@wkeComps/QueryEditor`
dir and renamed as `conn-man`.

This removes `data-migration` route as it will be a part of
the workspace page, rendered as a child component of the worksheet.