MXS-4430 Add `migr-creation-dlg` component
Clicking on `Data Migration` task will show this dialog. This is done
this way to have a consistent UI/UX because clicking on `Run Queries` task
also opens a dialog.

This dialog has an input allowing the users to define the name of
the migration task. For the initial release, rename feat won't
be implemented.