MXS-2709: Move connection string fixing into ODBC
This way any connections created via the driver do not end up using the
very much unwanted default behavior of some of the drivers. This should
make query editor usage much more easy as the documentation for the
drivers, especially the Postgres one, might be a bit hard to comprehend
at times.