MXS-4430 Don't show etl result status in snackbar
Showing result status for async queries would cause confusion.
e.g. If the users add a new worksheet and doing other stuff while waiting for
the ETL result, the snackbar will show in a wrong context.  Anyway,
tt's not necessary to show it at all because the status is already
shown in the `data-migration` worksheet.