MXS-2709: Fix PostgreSQL type conversions
The DATETIME type is a better match to the PostgreSQL TIMESTAMP type and
it has none of the automatic behavior that exists for the MariaDB
TIMESTAMP type.

The BYTEA type was not handled and can be converted into a LONGBLOB.

The identifier name given to pg_get_serial_sequence was also not
correctly formed: it lacked the schema name and the values were not
quoted.