MXS-4494: Fix binlogfilter query event filtering
The filtering appears to no longer work with newer MariaDB versions
which reject DDL transactions that are essentially empty. Replacing the
contents of the query events seems to solve the problem. Row events used
with row-based replication do not seem to suffer from this problem.

Added a test case into the binlogfilter test that reproduces the problem
and verifies that it is fixed by this commit.

Also updated the documentation to describe how query events are handled.