Prevent tidle expansion in copy_to_node
Newer versions of scp use the SFTP protocol which does not support file
globbing or tilde expansion to to the home directory. The result of this
was that when a directory was being copied over, the whole home
directory had its permissions modified which in in turn broke SSH and
made the VM inaccessible.

Removed the copy_to_node_legacy function as it is no longer needed.

Modified existing code to use explicit paths where needed.