MXS-4430 Fix `sql-editor` background color
When the `etl-editor` is rendered in a worksheet and the other worksheet
is rendering the Query Editor, the background color of the `etl-editor`
also applies to the Query Editor. Dynamically changing the colors or theme
won't work.

This uses css for changing the background color of monaco editor. The
only downside is that if the `monaco-editor` package changes the class
name of editor elements, this css won't work. Anyway, if that happens,
this would be just a background color issue, it won't have any real
effect on the users.