MXS-2709: Document catalog field for generic ETLs
The ODBC API requires a third value for the catalog. This is sometimes
implied by the connection string of the ODBC connection but the generic
ETL needs it for the catalog information functions: they require three
values to accurately locate the target table.

The generic ETL is currently limited to doing operations on tables in
the same catalog but if this turns out to be useful, it can be made into
an optional value for each table.