MXS-4430 Add button to change active connection
Before the introduction of the "Workspace", there was a dropdown
placed in the worksheet navigation bar. It shows the active
connection, "unbound" connections, and an option to create a new
connection. It was removed because it doesn't make sense to place
it there since the worksheet is now not only about the query editor
but also ETL. However, when a connection expires or when the users
want to create a new connection for the active worksheet, there
should be a way to do so.

- This adds a button to create a new connection and change connection.
- This stores target connection name in `meta.name` as SkySQL does
that because the server name in SkySQL is hidden from the user. To
avoid duplicated code, maxgui query editor will also do that.
- This allows creating connections to the same maxscale server, listener,
or server object and binding those to separate worksheets. There is no
technical reason to limit that ability except the fact that those
worksheets would have the same name. However, that could be handled
later by providing an input for defining connection name.