MXS-4430 Clean up previous connection after binding the new one
If there is a previous connection bound to the active worksheet,
it should be disconnected right after binding a new one. From the
user's point of view, there is no need to keep multiple connections
for a worksheet except for query tab connections as they are connecting
to the same target but under different connection ids. If the user
wants to connect to a different one while keeping the current
connection, they would simply add a new worksheet.

This also deletes obsolete unit tests.