MXS-4430 Handle show error message for 404 request sent by `queryHttp`
If the connections expires after calling `validateConns`, connections
data will still be stored in IndexedDB. As a result, when the users
navigate a way from the workspace or delete query editor worksheet
or disconnect ETL connections, it shouldn't show the error message
in the snackbar.
The error is supposed to be shown when the users run queries, start
ETl task on expired connections.