MXS-2709: Support some PostgreSQL internal types
These are merely added as a convenience as importing the internal
PostgreSQL tables serves as a good functional benchmark.

Fixed the capitalization in a few places, makes it easier to look at.

Forced the type names into lowercase when using
information_schema.columns as the source: some of the ARRAY types were
in uppercase for some reason which caused errors when they were inserted
into JSON columns.