MXS-4430 Stop spinning etl status icon for each table
If a table has `error` or `execution_time` field the icon should stop
spinning. If a table doesn't have one of those fields, it should spin
until etl task stops.

This also renames `isRunning` props of `etl-status-icon` component
to `spinning`. This component is used for both etl task status
in general and each table status, so `spinning` is a better name.