MXS-4430 Refactor `db-input` to `timeout-input`
If database type is `generic`, label for `db-input` turns into
`Catalog` and the required error message is also different. In
addition, the destination doesn't need this input. Therefore,
this refactors it into `timeout-input`.

This "DRY" `etl-src-conn` and `etl-dest-conn` to make it more
reusable. Later on, `etl-src-conn` will be used for QueryEditor.

This also hides `Connection String` input if the `Advanced` switch is
off.