Do not treat empty config values as errors
This allows setting the value of a list to empty resulting
in an empty list.

Other empty values are validated and cause an error in the same way
as an incorrect value does except for ParamString. This can cause
problems when a new configuration is created or a configuration
is changed, in any case the cause of the problem should be clear
to the user.