Fix REST-API and MaxCtrl unit tests
The tests assumed that localhost maps to the IPv4 loopback address but
this assumption is broken by Node.js 17 where the default order in which
the IPs are returned has changed. A more robust solution is to use the
exact IP address that MaxScale listens on in the tests: this is already
what both MaxScale and MaxCtrl do by default.