MXS-4430 Handle show a standalone button in the `Migration` result stage
If the etl task is running, the only possible action is to cancel, having
a manage dropdown would be a two steps process for the users and it's also
not immediately clear what action could be done. In addition, after the etl
task is done and if the connections already expires, delete is the only
action available to the user.