MXS-4430 Use `QUERY EDITOR` as the default name for QueryEditor worksheet
If connections bound to a worksheet expires, the Query Editor
worksheet was renamed to `WORKSHEET` which is the default name.
However, it is not suitable to use that name for QueryEditor anymore
as it's now used for `BlankWke`. `QUERY EDITOR` is a right name for
it as the worksheet is for QueryEditor.