Fix TestConnections::run_test
If the test leaked memory, the run_test() function did not treat it as
an error and the test passed. This happened as the `global_result`
variable was updated inside cleanup() which was done after the
assignment to `test_errors` was done. Simply replacing `test_errors`
with `global_result` fixes the problem.