MXS-4430 Disconnect the connection if it fails to reconnect
In practive, for QueryEditor, if a query receives a lost connection
error, connections of other query tabs may have been lost as well unless
the users manually send a `KILL thread_id` statement where the `thread_id`
belongs to the connection that receives the lost connection error.
Either way, showing `Reconnection attempt failed. Please delete the active
query tab and create a new one.` is not an optimal approach. If there
is only 1 query tab, deleting it won't delete the connection. So,
it's better to just disconnect all connections bound to the worksheet.

For ETL connections, only connection that can't be reconnected will
be disconnected.