MXS-4430 Add `EtlTaskTmp` model
Now ETL tasks are bound to worksheets meaning that ETL tasks
can be run parallelly. But the `etl_res` state in the `etlMem`
state module doesn't have ability to store result for multiple
etl tasks.

To make that works parallelly, the `EtlTaskTmp` model
is needed to easily store etl result for each etl task.

This adds the mentioned model and takes it into use but the
`etl_res` field in the `EtlTaskTmp` model hasn't been taken into
use.