MXS-4430 Refine workspace entities relationship
With the current entities relationship, the following entities
`QueryEditorTmp, SchemaSidebar, QueryTab` are on the same level
as `EtlTask` which is not sensible. A new entity called `QueryEditor`
should be created to sit in front of the mentioned entities and
placed on the same level as `EtlTask`.

This commit only renames `active_etl_task_id` to `etl_task_id`
because the `Worksheet` entity can only have 0 to 1 `EtlTask`, so
having `active` as the prefix isn't right.

The next commit will create `QueryEditor` entity.