MXS-4430 Add info link and update new image for workspace ORM diagram
This adds info link for ETL instruction and limitations.

This adds a tooltip for `Create mode` input instead of pointing
it to the documentation as the `create_mode` doesn't have a markdown
heading, so it can't be pointed. It'll be rendered as follows in the
tooltip:

`Modes for creating tables
 normal:  CREATE TABLE
 ignore:  CREATE TABLE IF NOT EXIST
 replace: CREATE OR REPLACE TABLE
`

This also removes obsolete one to one inverse FK as those are unused.