MXS-4430 Remove `etlMem` state module
States of this module can be moved to EtlTaskTmp and actions
can be then moved to `etlTasks`. By doing this, schema tree data
of multiple EtlTask worksheets can be cached in memory.

This also shortens the name of serveral EtlTask getters.
It is not necessary to explicitly name the getter with the 'ById'
suffix. If a getter is for retrieving an active object,
the active word in the name is sufficient to distinguish it.