MXS-4430 Handle expired target connection
If the source connection expires, this is handled already by
the `$queryHttp` axios error interceptor. However, if the target
connection expired right before preparing or starting etl,
the `target` field will be empty in the body request. As a result,
the API returns an error response with this message
"The `target` field is mandatory".
This is expected but it's meanless for the users using ETL GUI.

The `handleEtlCall` action should check source and target connections
before sending the request.