MXS-4430 Remove `axios_opts` state
This state is used for SkySQL to handle multiple MaxScales using within
a hidden worksheet. However, with the new UI/UX design. The workspace
won't work with that solution.
The workspace in SkySQL would have to enable the worksheet navigation,
so the users can open the Query Editor and perform Data Migration for
different Sky services

With that being enabled, the `$queryHttp` can not handle requests
sent by different worksheets at the same time because the `axios_opts`
state is designed to store one value of `baseUrl`. So if there are
two worksheets with two MaxScales running Data Migration, it ends
up getting the result of for one worksheet only and the other will
probably receive 404.

This refactors `api` methods to accept axios request config. In the
next commits, actions that call those api methods will get axios
request config from the new model called `WorksheetTmp`. This
model will be used for storing axios option in memory.