MXS-2709: Fix GEOMETRY handling
The PostGIS GEOMETRY type is a user-defined type. The data_type field in
information_schema wouldn't be displayed as GEOMETRY as it isn't a part
of PostgreSQL itself and it would only be shown in the udt_name field.

The insertion into MariaDB also must use the ST_GeomFromText
function. Otherwise MariaDB doesn't understand to convert it from the
WKT form into whatever is used internally.

Added this assumption about the type into the ETL limitations document.