Don't block MainWorker while processing a KILL
The KILL command processing requires cross-thread communication to find
out which servers have connections to the given sessions or to the given
user. The simplest way to achieve this is to serialize the execution
onto the MainWorker which can safely gather the information without a
risk of a deadlock.

The current code mistakenly used the Worker::call() function that blocks
for the duration of the call instead of the Worker::execute() function
which only posts the message without waiting for the response. As the
MainWorker is only used to gather the list of targets to KILL and no
further processing takes place, the move to Worker::execute() is safe.