MXS-4430 Use `request_config` in all state modules
All API requests made in the workspace are now using `request_config`.
Most of them using the active `request_config` which is the config for
the active worksheet because the requests are only made when worksheet
is active. However, to make ETL polling requests work parallelly among
worksheets, it uses the `getRequestConfigByEtlTaskId` getter.

This also shorten the names of some getters.