MXS-4430 Add customize slot and subtext for blank worksheet task cards
SkySQL has a slight difference in rendering task cards in the blank
worksheet. If there is no SkySQL services available, it should disable
that card, show that info as a subtitle in the card and render a
button which is used to create a new service.

This adds props to `workspace-ctr` and slots inside `blank-wke`
so SkySQL can customize for their needs.