MXS-4430 Decompose `migr-dlg`
To enable SkySQL users to choose a service for the Data Migration,
and set the value for request_config in the worksheet, the
component is now decomposed into `migr-delete-dlg` and `migr-create-dlg`.

When the library is packaged, `migr-create-dlg` is exported,
requiring SkySQL to import and customize it according to their needs.