MXS-4508: Don't use MYSQL_OPT_RECONNECT in monitors
This causes problems with error reporting and might end up causing the
session variables to not be set correctly: the character set was set
with an explicit query which is lost if not set with MYSQL_INIT_COMMAND.

The automatic reconnection could also cause mariadbmon to lose the locks
without losing the connection. This would cause confusing errors as the
connection was indeed lost but wasn't reported back up to the monitor.