MXS-4430 Make `etl-src-conn` component reusable
This renames the component to `odbc-form` and replace the title
section with `prepend` slot. Later on the QueryEditor can reuse
this `odbc-form` for creating odbc connection.

The `Database Type` input is currently required only by ETL, but it
would also benefit QueryEditor to know the database type being used.
Current functionalities and queries are only compatible with MariaDB.
For instance, the `information_schema.TABLES` in postgresql doesn't
have the following columns `CREATE_TIME, TABLE_ROWS, and ENGINE`,
so it can cause the query to fail.

This also moves connection creation fuction to `conn-dlg-ctr`.