Fix bugs found by Clang
The assignment from path to path ended up using the value being
initialized. This happened as the variable shadowed an existing one. A
similar bug in 23.02 was just recently found.

The move assignment for Configuration did nothing with the rvalues which
Clang warned about but GCC did not.

The extern "C" part in query_classifier.hh should not be needed as no C
code should include it. In addition, it would expose C++ constructs in
the API which would not work.