MXS-4430 Handle 401 while sending requests from the workspace
If the `queryHttp` axios instance catches a 401 response status
code, it should redirect to the login screen. However, the
workspace will be packaged as a library, so it can't directly
redirect the users to the login page.

This adds `is_session_alive` state and uses it for conditionally
redirecting the users to the login page in maxgui.
SkySQL can use this state to do something else. e.g. Show an
error message.

This splits the `logout` action into smaller actions.
If the `logout` action is called after receving the 401 status code,
there is no need to call action to disconnect workspace connections
because it won't succeed anyway. This also removes code to clear
data in IndexedDb as it's obsolete.