Log KILL failure warning in debug builds
The sr_basics test can end up failing if one of the DCBs for a session
that's about to be killed hasn't connected yet. To strengthen this
hypothesis, a warning is now logged in debug builds whenever this
happens.

Changed the sr_basics test to abort the test early if a failure is
observed.