MXS-4515: Stash fake EPOLLIN events when throttling
This prevents an infinite read loop from happening if the calling code
always generates a fake EPOLLIN event. This is not a problem in any
released version as the socket is always fully drained.

However in 23.02 this is a problem as the read size is limited to the
writeq size which is now much smaller than the TCP network buffer. This
caused an EPOLLIN event to be triggered on the DCB where the data was
read from due to the fact that only a part of the data was read. The
result of this was that it ended up ignoring the writeq throttling when
the triggered EPOLLIN events kept pushing more data to the client DCB.