MXS-4513 DRY chart options
The previous commit actually fixed the "jumping" bug.
The chart was detroyed and rerendered whenever the chart is pushed
with new data which caused the "jumping" bug. By using the built-in
`reactiveProp`, that bug was fixed.

This creates `base` mixin for charts and add `opts` watcher so whevener
the users change the "Next Refresh" value, the stream line chart will
rerender the chart.

This also takes the `base` mixin into use.