Adapt to connector-c v3.3 replication API changes
The buffer is now in the event instead of the main replication
handle. The semi-sync size also needs to be taken into account in order
for the event to be correctly processed. With these changes in place,
the only required step to take semi-sync into use is to add
`SET @rpl_semi_sync_slave=@@rpl_semi_sync_master_enabled` to the
initialization SQL.