MXS-3004: Enable semi-sync for replication
Now that the connector knows how to deal with semi-sync replication, it
can be enabled with one extra SQL command.

The connector doesn't seem to send a COM_REGISTER_SLAVE unless the
hostname and port are set in the options. With this, the replication
connections are visible in the SHOW SLAVE HOSTS output.

The test case verifies that the binlogrouter works when semi-sync
replication is enabled.