MXS-4523 Make preparations for protocol specific parser/classifier
In order to be able to support multiple protocols (passing through
MaxScale), the query classifier cannot be global but must be protocol
specific. However, to keep things simple the assumption is that all
protocols passing through MaxScale carry SQL.

The parsing capability is exposed by a class mxs::Parser from which
protocol specific parsers must be derived (it's beside the point that
at least initially qc_sqlite may be sufficient for e.g. Postgres).

When a filter or router session is created, the parser instance is
obtained via the client connection of the session and stored in a
member variable, which is accessible via a member function 'parser()'.
Thus, inside a filter or router session, any function call like

    qc_some_function(buffer);

is replaced with

    parser()->some_function(buffer);

Currently there is one derived class - MariaDBParser - where the
implementation simly calls existing qc_-functions. Later the the
concrete parser will be loaded by the protocol module.