MXS-4250 Clarify MonitorServer handling
The base class monitor now tells the implementation whenever the
configured servers change. The implementation is responsible for
generating MonitorServer objects and informing the base class
what the true monitored servers are.

Also, monitored servers are now communicated to services as an
array instead of 1-by-1.

XPandMonitor support not yet complete.