MXS-4452: Call upstream handleError directly
This makes it possible to propagate custom errors upwards which can then
be handled by an upper-level service, if possible. Currently this just
converts the implicit error propagation into an explicit one, similarly
to how routeQuery and clientReply work.

Moving the error propagation into RouterSession::handleError and not
making it a pure virtual function allows the empty handleError
definitions to be removed. Given that there's only two routers that
actually handle errors (readwritesplit and schemarouter), this is a
sensible approach.