Assert the state of a moved-from GWBUF
If the GWBUF is guaranteed to be in a consistent state after being moved
from, it simplifies the usage of it as a call to clear() might not be
needed after it was moved from. It already fulfilled this requirement so
this commit just adds assertions. However, since a call to std::move
doesn't actually guarantee that a move-assignment or move-construction
takes place, a call clear() is still required if the buffer is used
afterwards.

This commit also adds the operator bool for a buffer as a check whether
a buffer is empty is fairly common and a boolean operator makes it less
verbose.