Remove GWBUF pointers from readwritesplit
Most of the changes were about changing from mxs::Buffer to GWBUF
values. The rest of the code already used the buffers by value and thus
required little changes.

The Trx class now requires a copy-constructor and copy-assignment as the
GWBUF objects aren't copyable. Also added reference overloads for the
Checksum class.

A handful of functions in readwritesplit now take non-const references
to a GWBUF. This was done the keep the code changes minimal but these
can eventually be removed by restructuring the code. Mostly these
functions are related to the causal_reads functionality.