MXS-4452: Use protocol module to create error packets
This allows a generic human-readable error string to be used in the
handleError function.

The MXS_SESSION::kill now also takes a string instead of an GWBUF. This
simplifies its usage as it was used only to deliver custom error
messages to the client. Now that the protocol generates the errors, the
same code should work for all protocol types.

Currently the routers require custom MariaDB errors to be generated
which is why the make_error() function takes an extra argument in the
form of a MariaDB error number. For non-MariaDB protocols, this can be
ignored. Postgres, for example, has no concept of error numbers and the
SQLSTATE alone determines what kind of a problem it was.