MXS-4452: Use plaintext errors in handleError
This makes it easier to modify or amend the error in the router modules
when the error gets propagated upwards. The only real modifications
outside of the APIs were in readwritesplit where the log messages needed
to be converted into calls to mxb::string_printf.

Fixed some variable names to match the type declarations at the same
time that the error message type was changed.

Fixed the template arguments for mxs::Error::set where it assumed that
all of the iterators are of the same type.

This commit also changes the behavior of readwritesplit: previously a
custom error was sent to the client in the rare case that read retrying
was not enabled. The new behavior where the session is killed is
consistent with how readwritesplit handles other similar problems.