Remove modutil_create_query
Replaced the calls with mariadb::create_query which returns the GWBUF by
value.

This commit also removes the use of GWBUF pointers from the regexfilter
and commentfilter. The main use of the pointers was the query generation
and with the switch to mariadb::create_query, the code was trivial to
modify to the new version.