Make GWBUF(size_t) similar to gwbuf_alloc
Almost every use of this constructor did a call to write_complete() with
the exact same value right after the buffer was constructed. This
appeared to be a needless extra step and making it behave like
gwbuf_alloc will make converting code to the new style both easier and
more efficient.