Remove protocol specific code from GWBUF
Although the functions might've been useful in other protocols, they
were only used by one function in the MariaDB protocol and even there
replacing it with direct manipulation was more efficient.

The CDC protocol used them in a few places but modifying the write()
function overload to take a std::string_view also solves this problem.

The avrorouter also used it to zero-terminate the buffer but reserving
extra space and zeroing that out instead will achieve the same end
result.