MXS-4529: Handle StartupMessage responses
Once the StartupMessage has been sent the server responds with a stream
of packets until either a ReadyForQuery or ErrorResponse message is
received. After this point has been reached, the routing of packets
proceeds normally. For now, the data is written as-is and no result
tracking is done.

Any packets that are read in the write() function while the
authentication is ongoing are put into the backlog. The backlog is
emptied once authentication has completed.