Only allow strings in mxb::cat
Asserting that a std::string_view can be constructed from all the types
prevents the use of characters with it. This prevents the conversion of
intergers into characters which should nearly always be
accidental. Since the intended use-case is to concatenate
std::string_view-like objects and not characters, this is a convenient
way to avoid easy mistakes.