MXS-4532: List router and filter protocols
The set of protocols that a router or a filter supports must now be
defined by the router.

Added the protocols for all the routers and filters in MaxScale. Most of
them only support the MariaDB protocol but some of them do work with all
protocols.

There are modules, `kafkacdc` and `kafkaimporter`, that do not support
any protocols at all. Support for a module with no protocol at all would
make the conflict detection more complex and thus the two routers claim
to support all protocols. This should be OK as they never instantiate
the router sessions and any attempt to do so would fail.