MXS-4548 Fix the bug in common code
This is a prolific bug existing in multiple places. First
fix the bug in the common consume_comment() which is not
used everywhere (and specifically not where this bug
was found). The error was simply that a comment "/**/"
without whitespace inside was skipped over leading to the
loss of the rest of the line.