MXS-4532: Detect protocol incompatibilities
Mismatches in the protocols of listeners, filter and services are now
detected. The detection is done by recursing down into the targets of a
service. The resulting set intersection is computed during the request
as storing it would mean it would have to be kept up to date. As the
dependencies are checked whenever the configuration is modified, the
approach of choosing simplicity over performance in this case is
warranted.

The names of the network protocols are assumed to be static: if two
names do not compare equal they are not the same protocol, even if they
would compare equal when done case-insensitively. As long as the common
constants are used to define the set of supported protocols, this should
not cause problems.